buyv_backend/requirements.txt
absl-py==2.1.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.13.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncio==4.0.0
attrs==25.1.0
babel==2.17.0
backtrader==1.9.78.123
bcrypt==4.0.1
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
black==23.11.0
bleach==6.2.0
blinker==1.8.2
blobfile==3.0.0
Brotli==1.1.0
cachetools==5.5.2
ccxt==4.5.4
celery==5.3.4
Cerberus==1.3.5
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
coacd==1.0.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.10.6
crcmod==1.7
cryptography==43.0.1
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
etils==1.12.0
eval_type_backport==0.2.2
eventlet==0.36.1
executing==2.2.0
fastapi==0.115.0
fastjsonschema==2.21.1
fhir.resources==7.0.0
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
flake8==6.1.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Compress==1.14
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.7.0
Flask-Login==0.6.3
flask-marshmallow==1.3.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
fonttools==4.55.8
fqdn==1.5.1
fredapi==0.5.2
freetype-py==2.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
# Editable install with no version control (genesis-world==0.2.1)
-e c:\users\user\desktop\genesis\genesis
gevent==23.9.1
glfw==2.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
gotrue==2.9.1
graphviz==0.21
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.14.0
h2==4.3.0
h5py==3.13.0
hpack==4.1.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.7.1
httpx==0.24.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jwt==1.3.1
keras==3.9.0
kiwisolver==1.4.8
kombu==5.5.4
lazy_loader==0.4
libclang==18.1.1
libigl==2.5.1
limits==4.0.0
llama_models==0.1.2
llama_stack==0.1.2
llama_stack_client==0.1.2
llvmlite==0.44.0
lxml==5.3.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
MetaTrader5==5.0.5370
mistune==3.1.3
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
mujoco==3.2.5
multidict==6.1.0
multitasking==0.0.12
mypy==1.7.1
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openai==0.28.0
opencv-python==4.11.0.86
OpenEXR==3.3.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
ordered-set==4.1.0
oss2==2.19.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandas-ta==0.4.71b0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
peewee==3.18.2
Pillow==10.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.6.0
pooch==1.8.2
postgrest==0.13.2
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaml==25.1.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.11.0
pycodestyle==2.11.1
pycollada==0.9
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.21.0
pydantic==2.9.1
pydantic_core==2.23.3
pydicom==2.4.3
pyflakes==3.1.0
PyGEL3D==0.6.1
pyglet==2.1.3
pygltflib==1.16.0
Pygments==2.19.1
PyJWT==2.9.0
pymeshlab==2023.12.post2
PyMySQL==1.1.0
PyOpenGL==3.1.9
pyparsing==3.2.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.12.2
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2024.2
pyvista==0.44.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
qwen-api==1.4.16
realtime==1.0.6
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.1
scooby==0.10.0
screeninfo==0.8.1
seaborn==0.13.2
selenium==4.31.0
Send2Trash==1.8.3
sentry-sdk==1.38.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.32
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
stripe==14.1.0
structlog==24.4.0
supabase==2.3.0
supafunc==0.3.3
sympy==1.13.1
ta==0.11.0
tabulate==0.9.0
taichi==1.7.3
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
terminado==0.18.1
tetgen==0.6.4
tifffile==2025.2.18
tiktoken==0.8.0
timm==1.0.15
tinycss2==1.4.0
together==1.5.8
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.115
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.30.0
vine==5.1.0
vtk==9.3.1
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.45.1
whitenoise==6.6.0
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.3
yarl==1.18.3
yfinance==0.2.65
zipp==3.21.0
zope.event==5.1.1
zope.interface==7.2
